+++ matminer/__init__.py
@@ -1,3 +1,3 @@
-__version__ = '0.5.7'
+__version__ = '0.5.9'
 
 from matminer.figrecipes.plot import PlotlyFig