+++ matminer/featurizers/composition.py
@@ -254,6 +254,81 @@ class ElementProperty(BaseFeaturizer):
     def implementors(self):
         return ["Jiming Chen", "Logan Ward", "Anubhav Jain", "Alex Dunn"]
 
+class Meredig(BaseFeaturizer):
+    """
+    Class to calculate features as defined in Meredig et. al.
+
+    Features:
+        Atomic fraction of each of the first 103 elements, in order of atomic number.
+        17 statistics of elemental properties;
+            Mean atomic weight of constituent elements
+            Mean periodic table row and column number
+            Mean and range of atomic number
+            Mean and range of atomic radius
+            Mean and range of electronegativity
+            Mean number of valence electrons in each orbital
+            Fraction of total valence electrons in each orbital
+
+    """
+
+    def __init__(self):
+        self.data_source = MagpieData()
+
+        #The labels for statistics on element properties
+        self._element_property_feature_labels = ["mean AtomicWeight", "mean Column", "mean Row", "range Number", "mean Number",
+        "range AtomicRadius", "mean AtomicRadius", "range Electronegativity", "mean Electronegativity"]
+        # Initialize stats computer
+        self.pstats = PropertyStats()
+
+    def featurize(self, comp):
+        """
+        Get elemental property attributes
+
+        Args:
+            comp: Pymatgen composition object
+
+        Returns:
+            all_attributes: Specified property statistics of features
+        """
+
+        #First 103 features are element fractions, we can get these from the ElementFraction featurizer
+        element_fraction_features = ElementFraction().featurize(comp)
+
+
+        #Next 9 features are statistics on elemental properties
+        elements, fractions = zip(*comp.element_composition.items())
+        element_property_features = [0] * len(self._element_property_feature_labels)
+
+        for i,feat in enumerate(self._element_property_feature_labels):
+            stat = feat.split(" ")[0]
+            attr = " ".join(feat.split(" ")[1:])
+
+            elem_data = [self.data_source.get_elemental_property(e, attr) for e in elements]
+            element_property_features[i] = self.pstats.calc_stat(elem_data, stat, fractions)
+
+        #Final 8 features are statistics on valence orbitals, available from the ValenceOrbital featurizer
+        valence_orbital_features = ValenceOrbital(orbitals=("s", "p", "d", "f"), props=("avg", "frac")).featurize(comp)
+
+        return element_fraction_features+element_property_features+valence_orbital_features
+
+    def feature_labels(self):
+        #Since we have more features than just element fractions, append 'fraction' to element symbols for clarity
+        element_fraction_features = [e + " fraction" for e in ElementFraction().feature_labels()]
+        valence_orbital_features = ValenceOrbital().feature_labels()
+        return element_fraction_features+self._element_property_feature_labels+valence_orbital_features
+
+    def citations(self):
+        citation = [
+            "@article{meredig_agrawal_kirklin_saal_doak_thompson_zhang_choudhary_wolverton_2014, title={Combinatorial "
+            "screening for new materials in unconstrained composition space with machine learning}, "
+            "volume={89}, DOI={10.1103/PhysRevB.89.094104}, number={1}, journal={Physical "
+            "Review B}, author={B. Meredig, A. Agrawal, S. Kirklin, J. E. Saal, J. W. Doak, A. Thompson, "
+            "K. Zhang, A. Choudhary, and C. Wolverton}, year={2014}}"]
+        return citation
+
+    def implementors(self):
+        return ["Amalie Trewartha"]
+
 
 class CationProperty(ElementProperty):
     """
@@ -1079,12 +1154,12 @@ class Miedema(BaseFeaturizer):
             'all': same for ['fcc', 'bcc', 'hcp', 'no_latt']
             ['fcc', 'bcc']: fcc and bcc solid solutions
         data_source (str): source of dataset, default='Miedema'
-            'Miedema': 'Miedema.csv' placed in "matminer/utils/data_files/"
-            parameterized for 73 elements by Miedema et al. in 1980s, containing
+            'Miedema': 'Miedema.csv' placed in "matminer/utils/data_files/",
+            containing the following model parameters for 73 elements:
             'molar_volume', 'electron_density', 'electronegativity'
             'valence_electrons', 'a_const', 'R_const', 'H_trans'
             'compressibility', 'shear_modulus', 'melting_point'
-            'structural_stability'
+            'structural_stability'. Please see the references for details.
     Returns:
         (list of floats) Miedema formation enthalpies (eV/atom) for input
             struct_types:
@@ -1508,6 +1583,11 @@ class YangSolidSolution(BaseFeaturizer):
             for e2, f2 in zip(elements[:i], fractions):
                 enthalpy += f1 * f2 * self.dhf_mix.get_mixing_enthalpy(e1, e2)
         enthalpy *= 4
+
+        # Make sure the enthalpy is nonzero
+        #  The limit as dH->0 of omega is +\inf. A very small positive dH will approximate
+        #  this limit without causing issues with infinite features
+        enthalpy = max(1e-6, abs(enthalpy))
 
         return abs(mean_Tm * entropy / enthalpy)